refactored lib.rs
