+++ crypto_utils/src/crypto_core.rs
@@ -1,463 +1,33 @@
-use crate::errors::{AesError, PgpError};
-use aes_gcm::{
-    aead::{Aead, AeadCore, KeyInit},
-    Aes256Gcm, Key as Aes_Key, Nonce,
-};
-use anyhow::Result;
-use argon2::Argon2;
-use base64::{engine::general_purpose, Engine as _};
-use openpgp::{
-    armor::{Kind::Signature, Writer as ArmorWriter},
-    cert::{CertBuilder, CipherSuite},
-    crypto::KeyPair,
-    packet::{
-        key::{PublicParts, SecretParts, UnspecifiedRole},
-        Key,
-    },
-    parse::{
-        stream::{MessageStructure, *},
-        Parse,
-    },
-    policy::{Policy, StandardPolicy},
-    serialize::{
-        stream::{Message, *},
-        Marshal,
-    },
-    types::{HashAlgorithm, KeyFlags},
-    Cert,
-};
-use rand::rngs::OsRng;
-use sequoia_openpgp::serialize::stream::Encryptor;
+use crate::aes_utils;
+use crate::errors::AesError;
+use crate::pgp_utils;
+use openpgp::Cert;
+use sequoia_openpgp::parse::Parse;
 use sequoia_openpgp::{self as openpgp};
-use std::error::Error;
-use std::io::Write;
-use std::io::{self};
+// Re-export functions from the specialized modules
+pub use aes_utils::{
+    decrypt_with_aes, encrypt_certificate, encrypt_with_aes, generate_aes_key, get_salt_arr,
+};
+pub use pgp_utils::{
+    decrypt_text_pgp, encrypt_text_pgp, generate_certificate, generate_key_id, get_decryption_key,
+    get_public_key_armored, get_recipient, get_signing_keypair, hash_text_sha512, sign_message,
+    verify_signature,
+};
 
-pub fn generate_certificate(username: &str) -> Result<openpgp::Cert> {
-    println!("Generating certificate for user: {}", username);
-
-    let (cert, _revocation) = CertBuilder::new()
-        .add_userid(username)
-        .set_cipher_suite(CipherSuite::Cv25519)
-        .add_subkey(KeyFlags::empty().set_signing(), None, None)
-        .add_subkey(KeyFlags::empty().set_storage_encryption(), None, None)
-        .generate()?;
-
-    println!("Certificate generated. Fingerprint: {}", cert.fingerprint());
-
-    Ok(cert)
-}
-
-pub fn derive_key(password: &str, salt: &[u8; 16]) -> Result<[u8; 32], AesError> {
-    let mut output_key_material = [0u8; 32];
-    Argon2::default()
-        .hash_password_into(password.as_bytes(), salt, &mut output_key_material)
-        .map_err(|e| AesError::KeyDerivationError(e.to_string()))?;
-    Ok(output_key_material)
-}
-
-pub fn encrypt_certificate(
-    data: &[u8],
-    password: &str,
-    salt: &[u8; 16],
-) -> Result<String, Box<dyn Error>> {
-    let key = derive_key(password, salt)?;
-    let cipher = Aes256Gcm::new_from_slice(&key)
-        .map_err(|e| AesError::CipherCreationError(e.to_string()))?;
-
-    let nonce = Nonce::from_slice(&salt[..12]);
-    let encrypted_data = cipher
-        .encrypt(nonce, data)
-        .map_err(|e| AesError::EncryptionError(e.to_string()))?;
-
-    Ok(general_purpose::STANDARD.encode(encrypted_data))
-}
-
-pub fn get_salt_arr(salt_b64: &str) -> Result<[u8; 16], Box<dyn Error>> {
-    let salt = general_purpose::STANDARD.decode(salt_b64)?;
-
-    if salt.len() != 16 {
-        return Err("Invalid salt length".into());
-    }
-
-    salt.try_into()
-        .map_err(|_| "Failed to convert salt to array".into())
-}
+/// High-level function that combines AES and PGP operations to decrypt a certificate
+/// This bridges the gap between the two crypto modules
 pub fn decrypt_certificate(
     encrypted_cert_b64: &str,
     salt_b64: &str,
     password: &str,
 ) -> Result<Cert, AesError> {
-    let encrypted_data = general_purpose::STANDARD
-        .decode(encrypted_cert_b64)
-        .map_err(|e| AesError::Base64DecodeError(e.to_string()))?;
+    // Use AES utilities to decrypt the raw certificate data
+    let decrypted_data =
+        aes_utils::decrypt_certificate_data(encrypted_cert_b64, salt_b64, password)?;
 
-    let salt_array = get_salt_arr(salt_b64).unwrap();
-
-    let key = derive_key(password, &salt_array)?;
-
-    let cipher = Aes256Gcm::new_from_slice(&key)
-        .map_err(|e| AesError::CipherCreationError(e.to_string()))?;
-
-    let nonce = Nonce::from_slice(&salt_array[..12]);
-
-    let decrypted_data = cipher
-        .decrypt(nonce, encrypted_data.as_ref())
-        .map_err(|e| AesError::DecryptionError(e.to_string()))?;
-
+    // Parse the decrypted data into a PGP certificate
     let cert = Cert::from_bytes(&decrypted_data)
         .map_err(|e| AesError::CertificateParseError(e.to_string()))?;
 
     Ok(cert)
-}
-
-pub fn get_public_key_armored(cert: &openpgp::Cert) -> Result<String> {
-    let mut buf = Vec::new();
-    cert.armored().serialize(&mut buf)?;
-    Ok(String::from_utf8(buf)?)
-}
-pub fn encrypt_text_pgp(
-    recipient: &Key<PublicParts, UnspecifiedRole>,
-    text: &str,
-) -> Result<String, anyhow::Error> {
-    // Perform the encryption
-    let mut encrypted = Vec::new();
-    {
-        let message = Message::new(&mut encrypted);
-
-        let recipient_obj = openpgp::serialize::stream::Recipient::new(
-            openpgp::types::Features::empty(),
-            recipient.key_handle(),
-            recipient,
-        );
-
-        let message = Encryptor::for_recipients(message, vec![recipient_obj])
-            .build()
-            .map_err(|e| PgpError::EncryptorCreationError(e.to_string()))?;
-
-        let mut writer = LiteralWriter::new(message)
-            .build()
-            .map_err(|e| PgpError::LiteralWriterCreationError(e.to_string()))?;
-        writer.write_all(text.as_bytes())?;
-        writer
-            .finalize()
-            .map_err(|e| PgpError::FinalizationError(e.to_string()))?;
-    }
-
-    // Armor the encrypted data
-    let mut armored = Vec::new();
-    {
-        let mut writer = openpgp::armor::Writer::new(&mut armored, openpgp::armor::Kind::Message)
-            .map_err(|e| PgpError::ArmorWriterCreationError(e.to_string()))?;
-        writer.write_all(&encrypted)?;
-        writer
-            .finalize()
-            .map_err(|e| PgpError::FinalizationError(e.to_string()))?;
-    }
-
-    Ok(String::from_utf8(armored)?)
-}
-
-pub fn decrypt_text_pgp(
-    policy: &dyn Policy,
-    decrypt_key: &openpgp::packet::Key<SecretParts, UnspecifiedRole>,
-    ciphertext: &[u8],
-) -> Result<Vec<u8>, PgpError> {
-    let helper = DecryptionHelperStruct {
-        decrypt_key: decrypt_key.clone(),
-    };
-
-    let mut decryptor = DecryptorBuilder::from_bytes(ciphertext)
-        .map_err(|e| PgpError::DecryptorCreationError(e.to_string()))?
-        .with_policy(policy, None, helper)
-        .map_err(|e| PgpError::PolicyApplicationError(e.to_string()))?;
-
-    let mut plaintext = Vec::new();
-    io::copy(&mut decryptor, &mut plaintext)
-        .map_err(|e| PgpError::DecryptionError(e.to_string()))?;
-
-    Ok(plaintext)
-}
-
-impl VerificationHelper for DecryptionHelperStruct {
-    fn get_certs(
-        &mut self,
-        _ids: &[sequoia_openpgp::KeyHandle],
-    ) -> Result<Vec<sequoia_openpgp::Cert>> {
-        Ok(Vec::new())
-    }
-
-    fn check(&mut self, _structure: MessageStructure) -> Result<()> {
-        Ok(())
-    }
-}
-struct DecryptionHelperStruct {
-    decrypt_key: openpgp::packet::Key<
-        openpgp::packet::key::SecretParts,
-        openpgp::packet::key::UnspecifiedRole,
-    >,
-}
-impl DecryptionHelper for DecryptionHelperStruct {
-    fn decrypt(
-        &mut self,
-        pkesks: &[openpgp::packet::PKESK],
-        _skesks: &[openpgp::packet::SKESK],
-        sym_algo: Option<openpgp::types::SymmetricAlgorithm>,
-        decrypt: &mut dyn FnMut(
-            Option<openpgp::types::SymmetricAlgorithm>,
-            &openpgp::crypto::SessionKey,
-        ) -> bool,
-    ) -> openpgp::Result<Option<openpgp::Cert>> {
-        if pkesks.is_empty() {
-            return Err(anyhow::anyhow!("No PKESKs provided"));
-        }
-
-        let mut pair = KeyPair::from(
-            self.decrypt_key
-                .clone()
-                .into_keypair()
-                .map_err(|e| PgpError::KeyPairCreationError(e.to_string()))?,
-        );
-
-        for pkesk in pkesks {
-            if let Some((algo, session_key)) = pkesk.decrypt(&mut pair, sym_algo) {
-                if decrypt(algo, &session_key) {
-                    // Return None instead of a Fingerprint as we don't have a Cert
-                    return Ok(None);
-                }
-            }
-        }
-
-        // If we've reached this point, decryption failed for all PKESKs
-        Err(anyhow::anyhow!("Decryption failed for all provided PKESKs"))
-    }
-}
-pub fn hash_text_sha512(text: &str) -> Result<Vec<u8>, PgpError> {
-    let builder = HashAlgorithm::SHA512
-        .context()
-        .map_err(|e| PgpError::HashContextCreationError(e.to_string()))?;
-
-    let mut context = builder.for_digest();
-    context.update(text.as_bytes());
-
-    context
-        .into_digest()
-        .map_err(|e| PgpError::DigestComputationError(e.to_string()))
-}
-
-pub fn get_recipient(public_key: &str) -> Result<Key<PublicParts, UnspecifiedRole>, PgpError> {
-    // Parse the public key bytes into a Cert
-    let cert = Cert::from_bytes(&public_key.as_bytes())
-        .map_err(|e| PgpError::CertificateParseError(e.to_string()))?;
-
-    // Create a policy for key selection
-    let policy = &StandardPolicy::new();
-    // Find a suitable encryption key
-    cert.keys()
-        .with_policy(policy, None)
-        .supported()
-        .alive()
-        .revoked(false)
-        .for_storage_encryption()
-        .next()
-        .map(|key| key.key().clone())
-        .ok_or(PgpError::NoSuitableEncryptionKeyError)
-}
-
-pub fn get_signing_keypair(cert: &Cert) -> Result<KeyPair, PgpError> {
-    let policy = &StandardPolicy::new();
-    let signing_key = cert
-        .keys()
-        .with_policy(policy, None)
-        .for_signing()
-        .unencrypted_secret()
-        .next()
-        .ok_or(PgpError::NoSuitableSigningKeyError)?;
-
-    signing_key
-        .key()
-        .clone()
-        .into_keypair()
-        .map_err(|e| PgpError::KeyPairCreationError(e.to_string()))
-}
-
-pub fn get_decryption_key(
-    cert: &Cert,
-) -> Result<openpgp::packet::Key<SecretParts, UnspecifiedRole>, PgpError> {
-    let policy = &StandardPolicy::new();
-
-    cert.keys()
-        .unencrypted_secret()
-        .with_policy(policy, None)
-        .supported()
-        .alive()
-        .revoked(false)
-        .for_storage_encryption()
-        .next()
-        .map(|key| key.key().clone())
-        .ok_or(PgpError::NoSuitableDecryptionKeyError)
-}
-
-pub fn sign_message(keypair: &KeyPair, message: &str) -> Result<Vec<u8>, PgpError> {
-    let mut signature = Vec::new();
-    {
-        let message_writer = Message::new(&mut signature);
-        let signer = Signer::new(message_writer, keypair.clone())
-            .map_err(|e| PgpError::SignerCreationError(e.to_string()))?;
-
-        let mut signer = signer
-            .detached()
-            .build()
-            .map_err(|e| PgpError::SignerCreationError(e.to_string()))?;
-
-        signer
-            .write_all(message.as_bytes())
-            .map_err(|e| PgpError::MessageWriteError(e.to_string()))?;
-        signer
-            .finalize()
-            .map_err(|e| PgpError::SignatureFinalizationError(e.to_string()))?;
-    }
-
-    let mut armored_signature = Vec::new();
-    {
-        let mut armor_writer = ArmorWriter::new(&mut armored_signature, Signature)
-            .map_err(|e| PgpError::ArmorWriterCreationError(e.to_string()))?;
-        armor_writer
-            .write_all(&signature)
-            .map_err(|e| PgpError::SignatureWriteError(e.to_string()))?;
-        armor_writer
-            .finalize()
-            .map_err(|e| PgpError::ArmorFinalizationError(e.to_string()))?;
-    }
-
-    Ok(armored_signature)
-}
-pub fn generate_aes_key() -> Aes_Key<Aes256Gcm> {
-    Aes256Gcm::generate_key(OsRng)
-}
-
-pub fn encrypt_with_aes(key: &Aes_Key<Aes256Gcm>, plaintext: &str) -> Result<String, AesError> {
-    let cipher = Aes256Gcm::new(key);
-
-    let nonce = Aes256Gcm::generate_nonce(&mut OsRng); // 96-bits; unique per message
-    let ciphertext = cipher
-        .encrypt(&nonce, plaintext.as_bytes())
-        .map_err(|e| AesError::EncryptionError(e.to_string()))?;
-
-    let mut combined = nonce.to_vec();
-    combined.extend_from_slice(&ciphertext);
-
-    Ok(general_purpose::STANDARD.encode(combined))
-}
-
-pub fn decrypt_with_aes(key_bytes: &[u8], encoded: &str) -> Result<String, AesError> {
-    let key = Aes_Key::<Aes256Gcm>::from_slice(key_bytes);
-    let cipher = Aes256Gcm::new(key);
-
-    let decoded = general_purpose::STANDARD
-        .decode(encoded)
-        .map_err(|e| AesError::Base64DecodeError(e.to_string()))?;
-
-    if decoded.len() < 12 {
-        return Err(AesError::DecryptionError("Invalid ciphertext".to_string()));
-    }
-
-    let nonce = Nonce::from_slice(&decoded[..12]);
-    let ciphertext = &decoded[12..];
-
-    let plaintext = cipher
-        .decrypt(nonce, ciphertext)
-        .map_err(|e| AesError::DecryptionError(e.to_string()))?;
-
-    String::from_utf8(plaintext).map_err(|e| AesError::Utf8ConversionError(e.to_string()))
-}
-
-pub fn generate_key_id(public_key: &str) -> Result<String, Box<dyn Error>> {
-    let cert = Cert::from_bytes(public_key.as_bytes())?;
-    Ok(cert.fingerprint().to_hex().to_lowercase())
-}
-pub fn verify_signature(
-    public_key: &str,
-    message: &str,
-    signature_b64: &str,
-) -> Result<bool, PgpError> {
-    // Parse the public key
-    let cert = Cert::from_bytes(public_key.as_bytes())
-        .map_err(|e| PgpError::CertificateParseError(e.to_string()))?;
-
-    // Decode the base64 signature
-    let signature_bytes = general_purpose::STANDARD
-        .decode(signature_b64)
-        .map_err(|e| PgpError::Base64DecodeError(e.to_string()))?;
-
-    // Create a verification helper
-    let helper = SignatureVerificationHelper {
-        cert,
-        verification_successful: false,
-    };
-    let policy = &StandardPolicy::new();
-
-    // Create the detached verifier
-    let mut verifier = DetachedVerifierBuilder::from_bytes(&signature_bytes)
-        .map_err(|e| PgpError::VerifierCreationError(e.to_string()))?
-        .with_policy(policy, None, helper)
-        .map_err(|e| PgpError::PolicyApplicationError(e.to_string()))?;
-
-    // Verify against the message
-    verifier
-        .verify_bytes(message.as_bytes())
-        .map_err(|e| PgpError::VerificationError(e.to_string()))?;
-
-    // Get the verification result from the helper
-    Ok(verifier.helper_ref().verification_successful)
-}
-
-// Verification helper struct following the reference pattern
-struct SignatureVerificationHelper {
-    cert: Cert,
-    verification_successful: bool,
-}
-
-impl VerificationHelper for SignatureVerificationHelper {
-    fn get_certs(&mut self, _ids: &[openpgp::KeyHandle]) -> openpgp::Result<Vec<Cert>> {
-        Ok(vec![self.cert.clone()])
-    }
-
-    fn check(&mut self, structure: MessageStructure) -> openpgp::Result<()> {
-        // For detached signatures, we expect exactly one signature group layer
-        for (i, layer) in structure.into_iter().enumerate() {
-            match (i, layer) {
-                // We expect exactly one signature group at layer 0
-                (0, MessageLayer::SignatureGroup { results }) => {
-                    // Check if any signature verification succeeded
-                    for result in results {
-                        match result {
-                            Ok(_) => {
-                                self.verification_successful = true;
-                                return Ok(());
-                            }
-                            Err(_) => {
-                                // Continue checking other signatures
-                                continue;
-                            }
-                        }
-                    }
-                    // If we get here, no signatures were valid
-                    return Err(anyhow::anyhow!("All signature verifications failed"));
-                }
-                _ => {
-                    // For detached signatures, we should only have one signature group
-                    return Err(anyhow::anyhow!(
-                        "Unexpected message structure for detached signature"
-                    ));
-                }
-            }
-        }
-
-        if self.verification_successful {
-            Ok(())
-        } else {
-            Err(anyhow::anyhow!("No signature group found"))
-        }
-    }
 }
